move all code to DesktopAlertDialog with private access
